qml: set content height when orientation is horizontal in ListViewExt
We were already setting the content width, primarily for vertical
list view. However, there is no restriction as such that ListViewExt
can not be used in horizontal orientation.

By default, list view does not estimate content height when the
orientation is horizontal. However, in most cases if not all, the
delegate uses the content size to set its size. For example, in a
vertical list view, it is sensible that the delegate uses content
width to determine its width.

If the delegate does not set its width when orientation is vertical,
or height when orientation is horizontal based on content width and
height, respectively, then it is the view's responsibility to override
the content size.