qt: get rid of `OverlayMenu.qml`
`OverlayMenu` was being used to provide an
overlay menu within the play queue, which allowed
doing certain actions through keyboard such as
moving items.

However, the overlay menu functionality was
removed from the play queue with 75fdaaa6.

Now that there is no other use case for
`OverlayMenu`, and that there have not been
other cases where it has been used, it is
now time to get rid of it.