qml: handle padding properly in MusicArtistsAlbums
The row layout here can be considered the content item.
Padding of the root should be applied as margin for the
content item, like how `Control` behaves.

Currently, delegate is not positioned properly. The
leftmost side of the delegate should begin after the
safe margin.

This also enables taking care of the minimum width for
the row layout when there is padding.