module.rs: use separate variable for version
Using a separate variable for version allows matching the name of the C
define, ensuring grep would highlight both while making it clearer that
both needs to be updated. It also brings the benefit of removing the
null-character from the string itself as it can now be directly
concatenated when the code is generated, which ensures a future update
will not break the version by forgetting the null-character.