qt: fix undefined behavior when openvg scene graph backend is used
When OpenVG is used, the rectangle node provided does not offer
a material. At the same time, it is also not a null pointer but
it explicitly points to memory address "1" as a workaround done
by Qt.

Obviously, that memory address does not contain a real `QSGMaterial`
instance, so dereferencing the material causes segmentation fault.