qml: fix drag item no longer showing any image regression
Since 72f086c9, `RoundImage` has been used in `DragItem`.
However, it only works asynchronously and by the time
the input items are resolved, `RoundImage` instance(s)
are still in the progress of loading the image.

I added an intermediate state to wait for images to be
loaded. If the loading takes too much time, then the
timer is going to start the drag, so this change does
not make drag initalization synchronous.