qt: fix unable to drag a large collection of media
When `Drag.active` is unset, the FSM goes into inactive
state even though the input items are yet to be parsed.

Unlike loading done and loading failed states, inactive
state does not have an handling mechanism when the input
items are ready.

We should not fail the drag early if the input items can
not be collected until the drag operation finishes.
Ideally, this should be a synchronous operation, but it
can be asynchronously handled like this as well because
we can invalidate the drag if the user starts a new drag
operation before the previous operation is completed.