input: add input_GetItemDuration function to calculate input thread item duration
Reverts commit c86e464f and bf5b6d59

Add the input_GetItemDuration function, which calculates the duration of a
track item based on the start and stop times stored in the input thread's private
data. If the stop time is not set, the function considers the provided duration
as the stop time; otherwise, it calculates the duration using the start and stop times.

Call input_GetItemDuration for retrieving the updated duration of the current track
item. The calculated duration is then used to set the player's length accurately.