qml: provide the effective background color to the effect in TracksListPage
If "transparent" color is provided, fading edge can not do sub-pixel font
rendering. Translucent colors should be used only when absolutely necessary,
such as when window background needs to be shown for the backdrop blur
filter.

In this case, it already has a solid background due to `stackViewParent` in
`MainDisplay.qml`, that color should have been used instead.