qml: let `grabToImage()` do the off-screen rendering in DragItem
Trying to use `grabToImage()` on an invisible item
reportedly caused issues, possibly due to lack of
polish call or children just not getting ready as
soon as they are made visible off-screen.

With Qt 6, this does not seem to apply anymore.
Instead of rendering off-screen manually, it should
be a better idea to just let `grabToImage()` do that
itself.