filters: proxy: keep track of the proxied vars state
Since we now assume that the state of the filters parameters should be
stored in the vout (cf. 0be3282eb6d9a04a86160a74b780b2774e0252fa), the
state of the mapped variables should be kept between resets of the
filter chain.

The proxy functions used to create and destroy both the mappings and the
bound callbacks. In a perfect world, the mapping of the variables and
the binding of the callbacks should be decoupled. But for such a small
utility function, we can just stop destroying the state of the variables
between resets.

This patch also reworks the documentation of the modified functions
accordingly.

Refs #28769