qt: wait for Qt to initialize the dcomp device in `setupVoutWindow()`
This method requires the direct composition device
in order to create the video visual.

When this method is called, Qt might not have fully
initialized the interface as scene graph initialization
is done asynchronously.

In this case, simply wait for the initialization to
complete. If the initialization fails or takes too
much time to complete, then fail by returning false.

Co-authored-by: Pierre Lamot <pierre@videolabs.io>