qt: reduce playlist delegate vertical padding
Although having excessive padding makes it possible
to interact with the delegate using touch control
easier, it also causes the playlist to show less
items at once.

VLC is primarily a desktop application, although
we should care about touch support, we should not
optimize the design for touch.

Reducing the vertical padding from 6dp to 2dp seems
to be good compromise. It is still very much possible
to use touch, but also looks okay for a desktop
application.