qt: introduce ModelRecoveryAgent
This class can be used to restore a model if
it supports serialization to a former state,
recent state.

The intention is to use this for the playlist
list model to recover the playlist if the
application crashes, or unexpectedly closes.

A message box is shown when there is a
recoverable model found. The user has the
right to choose whether to recover or not.