sout: hls: keep up segmentation with stream time
Allow multiple segments to be output by the same pass.
This is mainly to handle parallel tracks that differ in segment sizes,
if one has its muxer outputting very small segments against a track with
the maximal segment size, the small segments must be output at a faster
rate.

This work is in preparation of the I-frame based segmentation where this
behavior will likely exist.