qt: setup startup for Darwin
Now that the interfaces are not run from the main thread directly, it's
possible to dispatch the QApplication startup and UI creation to the
main thread directly on Darwin. It matches the behaviour of the
interface on other platform that Darwin, where a thread is created to
handle the interface, except that the thread is already existing as the
main thread for Darwin.

It means that there won't need any specific code (except for Thread()
startup) to handle the specificities of the Qt interface startup on
Darwin anymore.

Ticket #27103 showed issues when leaving the interface with the
medialibrary, because the path releasing the medialibrary was not taken
correctly, but other issues could arise, including deadlocks.

Fixes #27103