qml: update playlist button click on video embed policy
when a video becomes embed the playlist is hidden, the playlist visibility
state is kept but the playlist was hidden, clicking the playlist button was
toggling the playlist visiblity. As a result, if the playlist was visible, it
required the user to click twice to show the playlist.

Now when the button is clicked with video embed, it will visually toggle the
playlist and update the playlist "visible" value to the visual state.