video_output: allow the AMD FRC filter as a static filter
"Interactive" filters can only do one output per input.
Only "static" filters are allowed to generate more frames, like the deinterlacers.

There is no way at this point to probe the module that will be used
if it will output many frames or not. And even if it could the order of
filters would need to be respected. We cannot dynamically add filters
on one chain or the other.