contrib: qtbase: disable building the unused direct2d platform plugin
Currently `direct2d` platform plugin is not used in
favor of `windows` platform plugin. So, it is not
necessary to build that platform plugin at the
moment. The matching feature for the platform plugin
is `direct2d1_1`.

The reason `direct2d` platform plugin is not used is
because it is not the default by Qt (as of Qt 6.8),
and it is not possible to statically link to both
`direct2d` and `windows` platform plugins due to
symbol clash, and `direct2d` menus are not supported
by Wine.

It should be noted that we still need `direct2d`
feature itself (`direct2d`), particularly because
of directwrite that is used for the font engine
and font database.