qt: synchronize the color schemes with Qt 6.8.0
Qt finally provides the ability to change the color
scheme. Although the default palette colors are
rarely used in Qt Quick, this is still an important
thing for the platform plugin.

This means that if the user opts in to a specific
scheme, things such as platform menus, windows,
window frames would adapt to the overridden scheme
rather than using the setting inherited from the
system.