clock: don't generate wait_sync_ref with clock-start
The conversion function might not have wait_sync_ref defined in the
next steps as it won't generate the wait_sync_ref itself, but choosing
the start_date requires transforming the current point.

The assertion on wait_sync_ref is removed given that it would be NULL
when using start_time.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>