qt: consider csd shadows when requesting blur behind effect
KWindowSystem's documentation states the following:

> Note that you will usually want to set the region to the
> shape of the window, excluding any shadow or halo.

I interpret this as, the requested blur region should not
encapsulate the area where the window shadows are drawn by
the client (application) because client shadows are part of
the window.

Historically this has not been a problem because the client
shadows are placed in the frame extents, and KWin did not
apply blur effect in the frame extents/margins. This does
not seem to be a documented behavior, and with KWin 6.2.0
it appears that the blur is applied in frame extents on X11.