clock: dump more metrics in traces
This adds the clock context id metric which allows tracking some of the
lipsync issues by checking every contexts are switching to the new
context reliably. In this patch, clock->context can be null, but in
future work it will never be NULL.

This also setup a new tag on traces exposed by TraceRender to separate
the metrics where ts is the current time and metrics where ts is a time
provided by the tracer user. The rationale is that some systems cannot
render non-monotonic metrics and the time at which the event appear is
more important than the value of the timestamp, which should be traced
separately.