qt: apply blur behind effect when the target window is activated
Currently, the effect is applied before the window is shown. Frankly,
there is no restriction from the compositors side (DWM, KWin) regarding
when the backdrop blur effect can be activated, and it was a courtesy
for the compositors to request the effect even before the window is shown
so that they could do the job potentially better.

For us, it does not matter much when the blur is applied. However,
with KWin Wayland it does not seem to be possible to use effects before
the window is shown. It is not clear if this is a KWin regression, but
since it does not matter, the blur can be requested when the window is
activated.