qml: accept all devices in Flickable tap handlers
Setting the accepted device to only mouse causes
problems. Sometimes, particularly on Wayland,
mouse does not identify as a mouse. At the same
time, track pad can not be used as well as it is
considered to be a different device (and not allowed
here).

This was used to prevent interfering with dragging
and flicking the view, which is a use case mainly
with touch screen. However, according to QTBUG-73035,
and commit 250053 [1], it appears that having a direct
tap handler within Flickable should not cause issues
with touch screen.

[1] https://codereview.qt-project.org/c/qt/qtdeclarative/+/250053