qt: fix CompositorDCompositionAcrylicSurface causing the interface window to be inactive
CompositorDCompositionAcrylicSurface creates a window itself, which makes the newly
shown interface window to be no longer active.

The interface window is just shown, it should be active. At the same time, the window
CompositorDCompositionAcrylicSurface creates is a dummy window, so it does not
make sense for it to request (and steal from the interface window) activation.