qt: handle non-`text/uri-list` multi-line text drop properly
A single URL can not contain multiple lines. If the input is
multi-line text, then it should be treated as multiple urls,
not a single url.

This is with accordance to `text/uri-list`, the main difference
is being more relaxed regarding the line ending. Unlike
`text/uri-list`, CRLF line ending is not mandatory and LF is
enough.