qt: fail early if QtQuick.Effects module presence does not match the expectation
If Qt version is at least 6.5.0, the application expects that QtQuick.Effects
module is available. If it is not available, the interface is going to encounter
error when a type provided by that module is used and end up being in broken state.

If we fail early, as we do with the other modules, then the interface will close
gracefully.

Qt always logs when an imported module is not found. The user should install the
required module using the package manager. This module is already provided with
the contrib Qt.