pulse: fail if the server is Pipewire
Some distributions allow installing both Pipewire and PulseAudio, leaving
Pipewire to manage only video devices.

On one hand, if we leave the Pipewire output with higher priority, then we
do not get access to any audio output device in this case. On the other hand,
if we simply probe the PulseAudio output first, then we will never use the
native Pipewire protocol, even if the PulseAudio service is not actually
running, since Pipewire emulates PulseAudio.

So this checks the name of the PulseAudio daemon, and uses the PulseAudio
protocol if and only if the server is not Pipewire.