qt: check for frameless window hint when probing Win11 22H2 window backdrop effect
Backdrop effect needs the interface window to be translucent, as we want the DWM
composited backdrop effect to pass through.

It is additionally required to use `Qt::FramelessWindowHint` for transparent
windows on Windows, otherwise artifacts may occur. Since this flag is explicitly
not used on Windows unlike other platforms, this change effectively means that
the native effect is not going to be used.

It is not clear why `Qt::FramelessWindowHint` is not used on Windows when CSD
is in use, because it is the canonical way of making a QWindow frameless.