opengl: vout_helper: store renderer as a filter
The filter provides access to the .sys pointer as well as the private
implementation from the core, whereas vlc_gl_renderer only references
the .sys pointer. Interacting with the filter in the filter engine
(vlc_gl_filters object) requires access to the vlc_gl_filter object
itself so store it instead of the vlc_gl_renderer.