libvlc: preparser: signal cancellation
This is in the continuity of aad952ea38f944eca55077d62be9d3d464236a3e.

The previous commit added an internal error code for preparsing
cancellation that wasn't handled in libvlc, the API stopped reporting
preparsing termination, which resulted in a medialibrary regression.

Previously, the cancellation was reported as a timeout event which is
misleading. This patch introduces a new status code to properly match
libvlc's internals.