qml: hide toolbar editor delegate until drop is completed
Currently, when `drop()` is called, the delegate is shown
in its original position with internal drag. This causes
the delegate to be shown for a momentary period of time
until it is removed from the view.

In order to prevent flickering, hide the delegate before
calling `drop()` and if the delegate is dropped in the same
view (intra-view move) then show it since the delegate
should not remain hidden in its new position.