CI: remove win64-llvm target
We already build Win64 with gcc and Win32 with LLVM.
This is redundant to test the code is building in the target and with the compiler.
There's also the UWP target that uses LLVM with the Win64 code.

We only need the nightly build which will only rebuild contribs when necessary.