caopengllayer: move opengl startup to dedicated function
The code will be moved towards a real OpenGL provider module in the
future. This commit starts writing the module like so by ensuring there
is a valid OpenGL startup function for the module.

Currently, the usage of the vout_display prevents from allocating the
views and everything from within the OpenOpenGL function, but they'll
get moved in later patches which will include setting up dedicated
callbacks for the rendering.