qml: do not use independent window for ToolTipExt on wayland
Although it seems to work, it is not reliable with Wayland.
With KWin Wayland compositor I often get the error:

qt.qpa.wayland: eglSwapBuffers failed with 0x300d, surface: 0x7be3c8512b70

I suspect that this occurs when the window is shown/hidden
before the fade animation ends. Since we don't control the
animation, and we have to support fast visibility changes
mainly (but not limited to) due to the pointing tool tip
that is used within the seek bar, we should go back to the
old approach with Wayland for the moment.

On Windows and X11, it so far has worked well for me.