qt: csdmenu: check QT_CONFIG(wayland) before using QNativeInterface
The Wayland parts are not available by default on the version compiled
by Homebrew, which makes the compilation break. On Qt version >= 6.7,
there is a QT_CONFIG(wayland)[^1] parameter which can be used to detect
whether the support is enabled or not.

Fatih found that before this version, Q_OS_UNIX was used in the code so
we can also detect that in the other cases.

[^1]: https://github.com/qt/qtbase/commit/a2ff329d2eb893ffae0d1cd18dd76bd494781295#diff-6ce1c6c83aa4b5db78a6748c8a9eec19620a10a2e2a495fd9053fdf391a9863aL4389

Co-authored-by: Fatih Uzunoglu <fuzun54@outlook.com>