opengl: sampler: add support for GL_EXT_YUV_target
GL_EXT_YUV_target[^1] provides a way to access an external texture
without applying a conversion to RGBA texture, which can be used to
supply our own matrices for conversion.

This can be used to workaround bogus RGBA conversion for some devices
and can help integrating better with deinterlacing, or in general plane
filtering, and some YUV pipeline like encoding.

[^1]: https://registry.khronos.org/OpenGL/extensions/EXT/EXT_YUV_target.txt