contrib: regenerate the meson contrib file to use local pathes with the tools
When using prebuilt contribs the native tools need to be rebuilt locally.
But the contrib.ini that is in the prebuilt tarball doesn't reflect the local pathes
for the native tools.

We can regenerate the contrib.ini files which will use the proper pathes
for these tools and the local contrib pathes for its constants.

Fixes #28896