Revert "contrib: regenerate the meson contrib file to use local pathes with the tools"
The dependencies to build meson-machinefile/contrib.ini builds some target
contribs packages when we only want to get the ini file alongside the native
tools.

This fails to build on Android because when using prebuilt contribs the
target compiler is not setup (config.mak) since it's not supposed to be used.

This reverts commit 51ff3023cc7f8e4a82b05a2df0c21dadb673c75b.