Makefile.am: fix installation regression
Automake doesn't generate automatic rules when a rule already exist in
the file, except if it cannot detect it. Adding the
`install-fooLTLIBRARIES: install-pkglibLTLIBRARIES` rule for audio
outputs, video outputs and vdpau broke the installation.

Adding `$(install-fooLTLIBRARIES): install-pkglibLTLIBRARIES` also makes
automake unhappy because of the variable name looking suspiciously close
to the target it generates, so a version without the dash is used for
the variable name in the rule. With this, the install target is
preserved but the dependency also exists in the output.