Makefile.am: fix installation regression
Automake doesn't generate automatic rules when a rule already exist in
the file, except if it cannot detect it. Adding the
`install-fooLTLIBRARIES: install-pkglibLTLIBRARIES` rule for audio
outputs, video outputs and vdpau broke the installation.

Adding `$(install-fooLTLIBRARIES): install-pkglibLTLIBRARIES` also makes
automake unhappy because of the variable name looking suspiciously close
to the target it generates, so a version without the dash is used for
the variable name in the rule. With this, the install target is
preserved but the dependency also exists in the output.

(cherry picked from commit 0b9f3f836f1c678e6d101f62545a8edaba16dcd0) (rebased)
rebased:
- the code around is slightly different