qml: hide source item in DynamicShadow only if it is actually visible
MultiEffect based DynamicShadow hides the source item because it
renders the source item itself again anyway. However, it does not
set itself invisible when it can not be used (software renderer).

The binding that makes the source item invisible checks if the effect
is visible. Here, we can simply bind the visibility to shader type
where the graphics backend needs to support RHI shader for the effect
to work. This is currently OpenGL, Vulkan, Metal, DirectX; so no
support for Software or OpenVG scene graph adaptations which are
not targeted by Qt RHI for obvious reasons.