qml: do not use linear filtering effect with non-RHI sg adaptation in ArtistTopBanner
Since blur effect is not available with the non-RHI scene graph adaptation (software,
and openvg adaptations), linear filtering was used here to approximate blur.

However with Qt 6, the texture does not seem to be aligned properly with the software
adaptation. This is most likely a Qt bug, but since the effect is not really important
for non-RHI adaptations, we can instead simply get rid of the effect and lower the
opacity instead.